requirements.txt
asttokens==2.0.5
beautifulsoup4==4.9.3
blis==0.7.4
catalogue==2.0.3
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
coverage==5.5
cycler==0.10.0
cymem==2.0.5
decorator==4.4.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0-py3-none-any.whl
executing==0.6.0
graphviz==0.16
html5lib==1.1
icecream==2.1.0
idna==2.10
Jinja2==2.11.3
joblib==1.0.1
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.4.1
murmurhash==1.0.5
networkx==2.5.1
nltk==3.6.2
numpy==1.20.2
packaging==20.9
pandas==1.2.4
pathy==0.5.2
Pillow==8.2.0
preshed==3.0.5
pydantic==1.7.3
Pygments==2.8.1
pyparsing==2.4.7
pytextrank==2.0.2
python-dateutil==2.8.1
pytz==2021.1
regex==2021.4.4
requests==2.25.1
scikit-learn==0.24.2
scipy==1.6.3
scraperapi-sdk==0.2.2
seaborn==0.11.1
six==1.15.0
sklearn==0.0
smart-open==3.0.0
soupsieve==2.2.1
spacy==3.0.6
spacy-legacy==3.0.5
srsly==2.4.1
style==1.1.0
thinc==8.0.3
threadpoolctl==2.1.0
tqdm==4.60.0
typer==0.3.2
update==0.0.1
urllib3==1.26.4
wasabi==0.8.2
webencodings==0.5.1
wordcloud==1.8.1
